Cherry picked Refactor of dimension list config from 1.21.1
Refactors the dimension list configuration to use a dedicated class
